+++ java/core/src/test/java/sleeper/core/record/process/status/ProcessRunsRecordsOutOfOrderTest.java
@@ -22,9 +22,9 @@ import java.time.Instant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.groups.Tuple.tuple;
+import static sleeper.core.record.process.status.ProcessRunsTestHelper.runsFromUpdates;
 import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.finishedStatus;
 import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.startedStatus;
-import static sleeper.core.record.process.status.TestProcessRuns.runsFromUpdates;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.TASK_ID_1;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.TASK_ID_2;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.onTask;
